backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==0.21.0
pydantic-settings==2.6.0
sqlalchemy==2.0.23  
alembic==1.13.0
psycopg2-binary==2.9.9

# LangGraph + LangChain stack
langgraph==0.2.33
langgraph-checkpoint-postgres==2.0.21
langchain==0.2.14
langchain-openai==0.1.22
langchain-core==0.2.41

# Bump httpx for clerk-backend-api
httpx==0.28.1
aiofiles==24.1.0
boto3==1.34.0
aioboto3==12.3.0
redis==5.0.1

# Clerk modern SDK
clerk-backend-api==3.3.0

slowapi==0.1.9
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
